UserService - register to create

+++ server-side/src/main/java/com/tick42/quicksilver/services/base/UserService.java
@@ -8,13 +8,13 @@ import com.tick42.quicksilver.models.specs.RegisterSpec;
 import java.util.List;
 
 public interface UserService {
-    UserModel setState(int id, String state);
+    UserModel setState(long id, String state);
 
     UserModel create(UserModel user);
 
     List<UserModel> findAll(String state);
 
-    UserModel findById(int id, UserDetails loggedUser);
+    UserModel findById(long id, UserDetails loggedUser);
 
     UserModel changePassword(NewPasswordSpec changePasswordSpec);
 }
